+++ tests/sunsynk/test_sunsynk.py
@@ -1,5 +1,5 @@
 """Test sunsynk library."""
-from typing import Sequence
+import logging
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
@@ -10,51 +10,51 @@ from sunsynk.pysunsynk import pySunsynk
 from sunsynk.rwsensors import NumberRWSensor
 from sunsynk.usunsynk import uSunsynk
 
-
-@pytest.fixture
-def sss() -> Sequence[Sunsynk]:
-    res: Sequence[Sunsynk] = []
-    if uSunsynk:
-        res.append(uSunsynk())
-    if pySunsynk:
-        res.append(pySunsynk())
-    return res
+_LOGGER = logging.getLogger(__name__)
 
 
 @pytest.mark.asyncio
-async def test_ss():
-    if pySunsynk:
-        ss = pySunsynk()
-        with pytest.raises(ConnectionError):
-            await ss.connect()
-
-
-@pytest.mark.asyncio
-async def test_ss_tcp():
-    if pySunsynk:
-        ss = pySunsynk()
-        ss.port = "127.0.0.1:502"
-        with pytest.raises(ConnectionError):
-            await ss.connect()
-
-
-@pytest.mark.asyncio
-async def test_ss_read(sss):
-    for ss in sss:
-        if uSunsynk:
-            ss = uSunsynk()
-            ss.client = AsyncMock()
-
-        if pySunsynk:
-            ss = pySunsynk()
-            ss.client = AsyncMock()
-
-
-@pytest.mark.asyncio
-async def test_ss_base_class():
-    ss = Sunsynk()
-    with pytest.raises(NotImplementedError):
+async def test_pyss():
+    ss = pySunsynk()
+    with pytest.raises(ConnectionError):
         await ss.connect()
+
+
+@pytest.mark.asyncio
+async def test_uss_schemes():
+    """Test url schemes for usunsynk.
+
+    umodbus only connects on read.
+    """
+    for port in ("serial:///dev/usb1", "tcp://127.0.0.1:502"):
+        ss = uSunsynk(port=port)
+        try:
+            await ss.connect()
+        except ModuleNotFoundError as err:  # not working on windows
+            _LOGGER.error("usunsynk could not connect to %s: %s", port, err)
+
+    for port in ("127.0.0.1:502", "xxx", "localhost"):
+        ss = uSunsynk(port=port)
+        with pytest.raises(ValueError):
+            await ss.connect()
+
+
+@pytest.mark.asyncio
+async def test_uss_sensor():
+    ss = uSunsynk(port="tcp://127.0.0.1:502")
+    await ss.connect()
+
+    rhr = ss.client.read_holding_registers = AsyncMock()
+
+    _LOGGER.warning("%s", dir(ss.client))
+    assert not rhr.called
+    await ss.read_holding_registers(1, 2)
+    assert rhr.called
+
+    wrr = ss.client.write_registers = AsyncMock()
+    assert not wrr.called
+    await ss.write_register(address=1, value=2)
+    assert wrr.called
 
 
 @pytest.mark.asyncio
@@ -64,6 +64,8 @@ async def test_ss_NotImplemented():
         await ss.connect()
     with pytest.raises(NotImplementedError):
         await ss.write_register(address=1, value=1)
+    with pytest.raises(NotImplementedError):
+        await ss.read_holding_registers(1, 1)
 
 
 @pytest.mark.asyncio
@@ -89,7 +91,7 @@ async def test_ss_write_sensor(rhr: MagicMock, wreg: MagicMock):
 
 @pytest.mark.asyncio
 @patch("sunsynk.Sunsynk.read_holding_registers")
-async def test_ss_read_sensor(rhr: MagicMock):
+async def test_ss_read_sensors(rhr: MagicMock):
     ss = Sunsynk()
     sen = NumberRWSensor((1,), "", min=1, max=10)
     rhr.return_value = (1,)
@@ -100,24 +102,3 @@ async def test_ss_read_sensor(rhr: MagicMock):
     rhr.side_effect = Exception("a")
     with pytest.raises(Exception):
         await ss.read_sensors([sen])
-
-
-@pytest.mark.asyncio
-async def test_ss_rhr():
-    ss = Sunsynk()
-    with pytest.raises(NotImplementedError):
-        await ss.read_holding_registers(1, 1)
-
-
-def test_update_sensor(caplog) -> None:
-    s = NumberRWSensor(60, "two", factor=0.1)
-    assert s.value is None
-    update_sensors([s], {})
-    assert s.value is None
-    update_sensors([s], {60: 10})
-    assert s.value == 1
-
-
-def test_bad_sensor(caplog) -> None:
-    NumberRWSensor((60, 1), "two", factor=0.1, bitmask=1)
-    assert "single register" in caplog.text